+++ liubility-typing-server/src/main/java/org/liubility/typing/server/service/ArticleService.java
@@ -12,5 +12,5 @@ import org.liubility.typing.server.domain.entity.Article;
 public interface ArticleService extends IService<Article> {
     Article getArticle(Article article);
 
-    Article getArticleById(Integer articleId);
+    Article getArticleById(Long articleId);
 }
