+++ liubility-typing-server/src/main/java/org/liubility/typing/server/service/impl/ArticleServiceImpl.java
@@ -25,7 +25,7 @@ public class ArticleServiceImpl extends ServiceImpl<ArticleMapper, Article> impl
     }
 
     @Override
-    public Article getArticleById(Integer articleId) {
+    public Article getArticleById(Long articleId) {
         Article article = getArticle(new Article(articleId));
         if(article == null){
             throw new LBRuntimeException("没有该文章");
